+++ packages/tokens/scripts/formats/css-vars.js
@@ -66,21 +66,27 @@ export const cssVariableFormatter = {
      * @type {string[]} List of variables that should be ignored for the brand 2025 theme
      */
     const BRAND2025_IGNORE_PATTERNS = [
-      // Needed for syn-range later on!
-      'input-border-color-offset',
-
       // Unknown component, skipping for now
       'typography-color-text-quiet',
       'typography-color-text-quiet-inverted',
       'color-primary-1000',
 
-      'interactive-background-color-action-hover',
-      'interactive-background-color-active',
+      'progress-track-readonly-color',
 
       'input-border-color-active',
       'input-icon-icon-clearable-color-active',
-      'input-readonly-background-opacity',
-      'input-readonly-background-border-color',
+      'interactive-background-color-active',
+      // Readonly color are needed later
+      'readonly-border-color',
+      'readonly-indicator-color',
+      'readonly-opacity-color',
+      'range-color-readonly',
+
+      // Font Style tokens are needed later
+      'body-2x-small-regular',
+      'body-2x-small-semibold',
+      'body-2x-small-bold',
+      'heading-4x-large',
     ].map(v => `${prefix}${v}`);
 
     /**